requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
rank_bm25
scikit-learn
numpy
pdfminer.six  